MLOps/requirements.txt
torch==2.6.0
typer==0.15.2
fastapi==0.115.12
uvicorn==0.34.2
torchvision
loguru
Pillow
kaggle
wandb
python-dotenv
fastapi==0.128.0
uvicorn[standard]==0.22.0
python-multipart
